Refactor deduction and income handling classes
- Updated `Archivos.h` to include `ConjuntoColaboradores.h`.
- Changed return type of `calcularSalario` in `Bonificaciones` to `double`.
- Modified `CCSS_Deduccion` constructor to accept `Colaborador*` and added new deduction constants.
- Enhanced `calcular` method in `CCSS_Deduccion` to use constants and check for null `colab`.
- Added getter methods for `SEM`, `IVM`, and `IPT` in `CCSS_Deduccion`.
- Removed `DeduccionConcretaBuilder` and implemented a new `Builder` pattern.
- Updated constructors of `DeduccionEmbargos`, `DeduccionMaternidad`, and `DeduccionRenta` to accept `Colaborador*`.
- Refactored `FactoryDeduccion` and `FactoryIngresos` to align with new parameters.
- Renamed `consola` to `Consola` and standardized method names.
- Updated `MenuArchivos`, `MenuColaboradores`, and `MenuPlanilla` for consistency.
- Enhanced `Modelo` with methods for file saving and loading.
- Introduced `DeduccionFyP` class for fixed or percentage deductions.
- Reorganized project files for improved clarity.

+++ proyecto c++/Bonificaciones.cpp
@@ -5,7 +5,7 @@ Bonificaciones::Bonificaciones(int t, float m, Colaborador* c) : Ingresos(c)
 	tipo = t;
 }
 
-float Bonificaciones::calcularSalario() const
+double Bonificaciones::calcularSalario() const
 {
 	if (tipo == 1) {
 		return colab->getSalarioBase();/*monto fijo*/
