+++ proyecto c++/HorasExtra.h
@@ -7,6 +7,6 @@ private:
 
 public:
 	HorasExtra(int h, Colaborador* c);
-	virtual float calcularSalario() const override;
+	virtual double calcularSalario() const override;
 };
