+++ proyecto c++/HorasOrdinarias.h
@@ -7,7 +7,7 @@ private:
     int horario;
 public:
     HorasOrdinarias(int h, Colaborador* c);
-    virtual float calcularSalario() const override;
+    virtual double calcularSalario() const override;
     
 };
